Big chunk of change in one commit to consolidate orphaned repo branch: commit messages from the source (top is least recent)
Change dispatching from ListenerMgr so that clients no longer need to have separate accept and handle functions. Now handle is the only function called by the manager and it returns true if the client handled the event.

Expand example 2 so that it scrolls more readily on a larger display which makes it better suited to working on the position of the status bar.
Make the status bar respond to scroll events. This seems to fix the status bar position in cases where the whole document, not just the editable, is scrolled (e.g. example2).
Make the FocusTracker generate scroll publications. Not necessarily the right place for this, but it'll do for now.

First go at making the toolbar viewport relative that even vaguely works. Should be ok on desktop Chrome. Doesn't handle all orientation changes correctly on iOS.

Give editables ids in example2 to make the example better for debugging.
Start re-working the focus tracker. This is a larger rework than previously as attempts to fix things round the edges haven't worked. This is a check point but seems to work in a number of cases where the previous version didn't work.

More changes for focus tracking.

Changes for focus tracking. This version seems to make things a lot better on both desktop Chrome and iOS. However it's still very untested and does too much work on platforms that support selection change events.

Try to make the selection tracking only fire when necessary by only using the selectionchange event when it is implemented.

Small improvements on selection tracking.
Make sure that information overlays (e.g. help) are handled in the same way as plugins. This means switching off key handling and making sure that auto save is also switched off when overlays are active. Previously auto save could have saved off the information overlay as part of the auto save.